feat: modularize ros2_graph_snapshot with new component detection and graph utilities
- Refactored `ros2_graph_snapshot.py` to utilize a modular structure, enhancing maintainability and functionality.
- Introduced new modules in `lib/snapshot` for component detection, graph management, and parameter collection, improving the overall architecture.
- Added `components.py` for querying composable node containers, `graph.py` for node graph information, and `params.py` for parameter handling.
- Enhanced process discovery capabilities in `proc.py`, allowing for detailed mapping of ROS 2 nodes to their respective processes and components.

These changes significantly improve the topology analyzer's ability to capture and report on ROS 2 system configurations, facilitating better analysis and insights.
